+++ CHANGELOG.md
@@ -4,8 +4,9 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ### Bugfixes
 
-* Fix range issue for `dgr`
+* Fix range computation for `dgr`
 * Fix IE bug occuring when using step count > 1 and `max` on loop counter
+* Fix argument check for `seq` for programs with offset
 
 ## v25.6.27